+++ lib/gds_api/test_helpers/calendars.rb
@@ -1,10 +1,6 @@
-require "gds_api/test_helpers/alias_deprecated"
-
 module GdsApi
   module TestHelpers
     module Calendars
-      extend AliasDeprecated
-
       def calendars_endpoint(in_division: nil)
         endpoint = "#{Plek.new.website_root}/bank-holidays"
         endpoint += "/#{in_division}" unless in_division.nil?
@@ -55,11 +51,6 @@ module GdsApi
       def stub_calendars_has_a_bank_holiday_on(date, in_division: nil)
         stub_calendars_has_bank_holidays_on([date], in_division: in_division)
       end
-
-      # Aliases for DEPRECATED methods
-      alias_deprecated :calendars_has_no_bank_holidays, :stub_calendars_has_no_bank_holidays
-      alias_deprecated :calendars_has_bank_holidays_on, :stub_calendars_has_bank_holidays_on
-      alias_deprecated :calendars_has_a_bank_holiday_on, :stub_calendars_has_a_bank_holiday_on
     end
   end
 end