+++ lib/gds_api/test_helpers/content_store.rb
@@ -1,4 +1,3 @@
-require "gds_api/test_helpers/alias_deprecated"
 require "gds_api/test_helpers/json_client_helper"
 require "gds_api/test_helpers/content_item_helpers"
 require "json"
@@ -6,7 +5,6 @@ require "json"
 module GdsApi
   module TestHelpers
     module ContentStore
-      extend AliasDeprecated
       include ContentItemHelpers
 
       def content_store_endpoint(draft = false)
@@ -93,13 +91,6 @@ module GdsApi
 
         stub_request(:get, url).to_return(body: body)
       end
-
-      # Aliases for DEPRECATED methods
-      alias_deprecated :content_store_has_item, :stub_content_store_has_item
-      alias_deprecated :content_store_does_not_have_item, :stub_content_store_does_not_have_item
-      alias_deprecated :content_store_has_gone_item, :stub_content_store_has_gone_item
-      alias_deprecated :content_store_isnt_available, :stub_content_store_isnt_available
-      alias_deprecated :content_store_has_incoming_links, :stub_content_store_has_incoming_links
     end
   end
 end