+++ lib/gds_api/test_helpers/support.rb
@@ -1,9 +1,6 @@
-require "gds_api/test_helpers/alias_deprecated"
-
 module GdsApi
   module TestHelpers
     module Support
-      extend AliasDeprecated
       SUPPORT_ENDPOINT = Plek.current.find("support")
 
       def stub_support_foi_request_creation(request_details = nil)
@@ -21,8 +18,6 @@ module GdsApi
       def stub_support_isnt_available
         stub_request(:post, /#{SUPPORT_ENDPOINT}\/.*/).to_return(status: 503)
       end
-
-      alias_deprecated :support_isnt_available, :stub_support_isnt_available
     end
   end
 end