+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+* BREAKING: Remove deprecated test helper methods that weren't prefixed with `stub_`
+* BREAKING: Remove `GdsApi::PublishingApiV2` use `GdsApi::PublishingApi`
+
 # 63.6.0
 
 * Remove request body parameters from error reporting payload.